pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["Kristina Pianykh <kristinavrnrus@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.96.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
httpx = "^0.24.1"
sense-hat = "^2.4.0"
pyyaml = "^6.0"
pysiddhi = "^5.1.0"

[tool.poetry.dev-dependencies]
poethepoet = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
start = "uvicorn server.http_server:app --reload --host 0.0.0.0"
